Refactor StageId (#242)
