+++ src/stages/mod.rs
@@ -13,16 +13,16 @@ mod total_gas_index;
 mod total_tx_index;
 mod tx_lookup;
 
-pub use block_hashes::BlockHashes;
-pub use bodies::BodyDownload;
-pub use call_trace_index::CallTraceIndex;
-pub use execution::Execution;
-pub use finish::Finish;
-pub use hashstate::{promote_clean_accounts, promote_clean_storage, HashState};
-pub use headers::HeaderDownload;
-pub use history_index::{AccountHistoryIndex, StorageHistoryIndex};
-pub use interhashes::Interhashes;
-pub use sender_recovery::SenderRecovery;
-pub use total_gas_index::TotalGasIndex;
-pub use total_tx_index::TotalTxIndex;
-pub use tx_lookup::TxLookup;
+pub use block_hashes::*;
+pub use bodies::*;
+pub use call_trace_index::*;
+pub use execution::*;
+pub use finish::*;
+pub use hashstate::*;
+pub use headers::*;
+pub use history_index::*;
+pub use interhashes::*;
+pub use sender_recovery::*;
+pub use total_gas_index::*;
+pub use total_tx_index::*;
+pub use tx_lookup::*;
